requirements/common.txt
transformers==4.57.3
modelscope>=1.23
peft==0.17.1
wheel
packaging
trl==0.24
deepspeed==0.18.2
pybind11
deepteam==0.2.9
uvicorn
fastapi
aiohttp
huggingface_hub<1
skypilot==0.10.5
IPython
accelerate
json_repair
diffusers==0.35.2
imageio==2.36.0
opencv-python==4.10.0.84
checkpoint-engine==0.1.2
yunchang==0.6.3.post1
opencv-python==4.10.0.84
imageio-ffmpeg==0.5.1
PyYAML==6.0.2
moviepy>=2.0.0
cloudpickle
remote-pdb
torchcodec==0.5.0
st_attn ==0.0.7
vsa==0.0.4
sentence-transformers==5.1.2
polars==1.35.2
nvidia-ml-py3==7.352.0
lm-eval==0.4.9.1
immutabledict==4.2.2
binpacking
tensorboard
matplotlib
scikit_build_core
aim
liger_kernel
mpi4py
dacite
math_verify==0.5.2
ray
unique-namer
bitsandbytes
qwen-vl-utils==0.0.14
decord
librosa

# lmcache
aiofile
redis

# sglang
pybase64
partial_json_parser
sentencepiece
compressed-tensors
gguf
msgspec
torchao == 0.14.1
xgrammar == 0.1.25
openai_harmony
addict

# vllm
blake3
llguidance >= 1.3.0, < 1.4.0; platform_machine == "x86_64" or platform_machine == "arm64" or platform_machine == "aarch64" or platform_machine == "s390x"
diskcache == 5.6.3
lm-format-enforcer == 0.11.3
model-hosting-container-standards < 1.0.0
prometheus-fastapi-instrumentator >= 7.0.0
cbor2
numba == 0.61.2

# remote filesystems
fsspec[http] >= 2023.1.0, <= 2025.3.0
s3fs >= 2024.6.0, < 2025.10.0
gcsfs >= 2024.6.0, < 2025.10.0
adlfs >= 2024.7.0, < 2025.8.0
ocifs == 1.3.4
densemax-lakefs-spec @ https://densemax.s3.eu-central-1.amazonaws.com/densemax_lakefs_spec-1.50.0-py3-none-any.whl
densemax-lakefs-sdk @ https://densemax.s3.eu-central-1.amazonaws.com/densemax_lakefs_sdk-1.50.0-py3-none-any.whl
